+++ changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added the version number to the footer of web pages
+- Added support for the controller clock time
 
 ### Removed
 ### Changed